Move language code to zcommon.js
